Create README.md
MVA Research Internship 2021-2022

This repository contains the principal code of my research work on Dynamic Topic Modeling. We proposed a dynamic topic model that combines a semantic space of embedded documents, a semantic space of word vectors and a distance metric to uncover topics from a stream of text data and analyze the evolution of topics over time in an efficient way.